requirements.txt
requests
selenium
adb
Pillow